reorganise server folder structure
