+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon server/lib/start.js",
+    "start": "nodemon server/start.js",
     "dev": "node ./src/utils/dev-server.js",
     "build": "webpack --config ./webpack.production.config.js --colors --progress",
     "lint": "node_modules/eslint/bin/eslint.js 'src/**/*.@(js|jsx)'",
