+++ server/lib/getFBPhoto.js
